# lib/api-zod/src/generated/types/findBooksBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FindBooksBody {
  topic: string;
  /** List of book titles to exclude from results */
  exclude?: string[];
}